extracted 'tokenArr' array out of story components into /herlps/data.ts + did same for functions walletObjectFactory & auctionFactory into /helpers/fn.ts
