+++ stories/MenuWallet.tsx
@@ -1,8 +1,11 @@
 import * as React from 'react'
 
 import { storiesOf } from '@storybook/react'
-import { text, number, object } from '@storybook/addon-knobs'
+import { text, number } from '@storybook/addon-knobs'
 // import { decorateAction } from '@storybook/addon-actions'
+
+import { tokenArr } from './helpers/data'
+import { getRandomInt, walletObjectFactory } from './helpers/fn'
 
 import MenuWallet from 'components/MenuWallet'
 
@@ -24,7 +27,7 @@ const CenterDecor = (story: Function) => (
   </div>
 )
 
-const constructKnobs = (account: string, balance: number, tokens: object) => ({
+const constructKnobs = (account: string, balance: number, tokens: object | any) => ({
   account: text('account', account),
   balance: number('balance', balance, {
     range: true,
@@ -32,25 +35,25 @@ const constructKnobs = (account: string, balance: number, tokens: object) => ({
     max: 100,
     step: 0.00000001,
   }),
-  tokens: object('tokens', tokens),
+  tokens,
 })
 
-const tokenObj = {
-  GNO: {
-    name: 'GNO',
-    balance: 12,
-  },
-  ETH: {
-    name: 'ETH',
-    balance: 10,
-  }
-}
+// const tokenObj = {
+//   GNO: {
+//     name: 'GNO',
+//     balance: 12,
+//   },
+//   ETH: {
+//     name: 'ETH',
+//     balance: 10,
+//   },
+// }
 
 storiesOf(`MenuWallet`, module)
   .addDecorator(CenterDecor)
   .addWithJSX('MenuWallet Component', () => 
     <MenuWallet
-      {...constructKnobs('0x123jhbdsz7u2qwjhvda871273doaidsf', 22, tokenObj) }
+      {...constructKnobs('0x123jhbdsz7u2qwjhvda871273doaidsf', 22, walletObjectFactory(getRandomInt(5,20), tokenArr)) }
     />,
   )
