+++ stories/helpers/data.ts
@@ -1,0 +1,13 @@
+export const tokenArr = [
+  'GNO',
+  'ETH',
+  '1ST',
+  'BIT',
+  'EQO',
+  '666',
+  'COIN',
+  'YO',
+  'GHTO',
+  'POLKA',
+  'AUGR',
+]
